+++ crates/zng-unit/src/lib.rs
@@ -22,6 +22,8 @@ mod side_offsets;
 mod time;
 mod transform;
 
+use std::fmt;
+
 #[doc(no_inline)]
 pub use euclid;
 
@@ -37,3 +39,51 @@ pub use px_dip::*;
 pub use side_offsets::*;
 pub use time::*;
 pub use transform::*;
+
+pub(crate) fn parse_suffix<T: std::str::FromStr>(mut s: &str, suffixes: &[&'static str]) -> Result<T, <T as std::str::FromStr>::Err> {
+    for suffix in suffixes {
+        if let Some(f) = s.strip_suffix(suffix) {
+            s = f;
+            break;
+        }
+    }
+    s.parse()
+}
+
+/// An error which can be returned when parsing an type composed of integers.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum ParseIntCompositeError {
+    /// Color component parse error.
+    Component(std::num::ParseIntError),
+    /// Missing color component.
+    MissingComponent,
+    /// Extra color component.
+    ExtraComponent,
+    /// Unexpected char.
+    UnknownFormat,
+}
+impl fmt::Display for ParseIntCompositeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ParseIntCompositeError::Component(e) => write!(f, "error parsing component, {e}"),
+            ParseIntCompositeError::MissingComponent => write!(f, "missing component"),
+            ParseIntCompositeError::ExtraComponent => write!(f, "extra component"),
+            ParseIntCompositeError::UnknownFormat => write!(f, "unknown format"),
+        }
+    }
+}
+impl std::error::Error for ParseIntCompositeError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        if let ParseIntCompositeError::Component(e) = self {
+            Some(e)
+        } else {
+            None
+        }
+    }
+}
+impl From<std::num::ParseIntError> for ParseIntCompositeError {
+    fn from(value: std::num::ParseIntError) -> Self {
+        ParseIntCompositeError::Component(value)
+    }
+}